Allow undefined in terminal env map, unify interfaces
Note that this adds undefined as an option to TerminalOptions.env which could end up
causing compile errors, but they will be more correct and could reveal a bug in an
extension since some key could return undefined.

Fixes #119046
